Update patch set 1
Patch Set 1:

I feel this is needed

value was calculated by:

1350/1200*[1200mhz current]

I think this is used to estimate powerconsumption better. In reality I believe these values don't do much but I think they are used for deciding how much is used on the battery for creating the report in Settings as we know it.

Patch-set: 1