Update patch set 1
Patch Set 1:

In that file aren't all necessary vendor files to AOSP ROMs. You can see it in libwvdrm_L1.so, libwvm.so, ibWVStreamControlAPI_L1.so and libdrmwvmplugin.so. They are used for DRM APIs and are in 4.2.2 too.

File libwvdrmengine.so is new for 4.3, and I suppose that it can be useful for futures uses.

Patch-set: 1